+++ traefik.tf
@@ -52,43 +52,9 @@ resource "kubernetes_network_policy" "traefik_allow_ingress" {
           }
         }
       }
-    }
-
-    policy_types = ["Ingress"]
-  }
-}
-
-resource "kubernetes_network_policy" "traefik_allow_cert_manager" {
-  count = var.network_policies.cert_manager.enabled ? 1 : 0
-
-  metadata {
-    name      = "${module.generic.namespace}-allow-cert-manager"
-    namespace = module.generic.namespace
-  }
-
-  spec {
-    pod_selector {
-      match_labels = {
-        "app.kubernetes.io/name" = "traefik"
-      }
-    }
-
-    ingress {
-      ports {
-        port     = 8000
-        protocol = "TCP"
-      }
 
       from {
         namespace_selector {
-          match_labels = {
-            "kubernetes.io/metadata.name" = var.network_policies.cert_manager.namespace
-          }
-        }
-        pod_selector {
-          match_labels = {
-            "app.kubernetes.io/name" = var.network_policies.cert_manager.name
-          }
         }
       }
     }