+++ variables.tf
@@ -44,20 +44,18 @@ variable "namespace" {
 
 variable "network_policies" {
   default = {
-    enabled       = true
-    ingress_cidrs = ["0.0.0.0/0"]
+    allow_ingress_enabled    = true
+    allow_monitoring_enabled = false
+    allow_namespace_enabled  = true
+    default_deny_enabled     = true
+    ingress_cidrs            = ["0.0.0.0/0"]
   }
-  description = "Traefik network policy customization"
+  description = "Traefik network policies customization"
   type = object({
     allow_ingress_enabled    = optional(bool, true)
     allow_monitoring_enabled = optional(bool, false)
     allow_namespace_enabled  = optional(bool, true)
-    cert_manager = optional(object({
-      enabled   = optional(bool, false)
-      name      = optional(string, "cert-manager")
-      namespace = optional(string, "cert-manager")
-    }), {})
-    default_deny_enabled = optional(bool, true)
-    ingress_cidrs        = optional(list(string), ["0.0.0.0/0"])
+    default_deny_enabled     = optional(bool, true)
+    ingress_cidrs            = optional(list(string), ["0.0.0.0/0"])
   })
 }